+++ StaticFields/headers/Model/Interpolate.h
@@ -5,7 +5,7 @@
 #ifndef INTERPOLATE_H
 #define INTERPOLATE_H
 //Helper function interpolating two vectors, fraction gives the distance between the vectors with 0 being at vector a and 1 at vector b
-Vector interpolateVec(Vector a, Vector b, long double fraction);
+Vector interpolateVec(Vector a, Vector b, double fraction);
 
 /**
  * Interpolates (Triliniarly) a field around a given vector position
