+++ StaticFields/headers/Model/Vector.h
@@ -3,9 +3,9 @@
 
 //standard 3D vectors
 typedef struct Vector {
-    long double x;
-    long double y;
-    long double z;
+    double x;
+    double y;
+    double z;
 } Vector;
 
 typedef enum dimensions {
@@ -18,16 +18,16 @@ typedef enum dimensions {
 Vector vecSum(Vector a, Vector b);
 
 //scalar multiplication
-Vector scalarMult(Vector a, long double mult);
+Vector scalarMult(Vector a, double mult);
 
 //same as scalar multiplication, but with division instead (1/divisor) * a
-Vector scalarDiv(Vector a, long double divisor);
+Vector scalarDiv(Vector a, double divisor);
 
 //creates a vector of all 0s
 Vector zeroVector();
 
 //gets vector value of given dimension
-long double getVecValue(Vector a, dimension dim);
+double getVecValue(Vector a, dimension dim);
 
 //subtraction of vector a and b (a-b)
 Vector vecMin(Vector a, Vector b);
